Moved keypad config into own file
