+++ lib/flutter_screen_lock.dart
@@ -1,6 +1,7 @@
 library flutter_screen_lock;
 
-export 'src/configurations/input_button_config.dart';
+export 'src/configurations/styled_input_button_config.dart';
+export 'src/configurations/key_pad_config.dart';
 export 'src/configurations/screen_lock_config.dart';
 export 'src/configurations/secret_config.dart';
 export 'src/configurations/secrets_config.dart';
